backend/demo/requirements.txt
requests
base58
cryptography
pyjwt
flatten_json
pyzbar
pillow